Replace prototype_parity test suite with s3api_verification test suite.
